+++ packages/architect/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @echojs-ecosystem/architect
 
+## 0.3.0
+
+### Minor Changes
+
+- update version v0.3.0
+
 ## 0.2.0
 
 ### Minor Changes