+++ packages/form/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @echojs-ecosystem/form
 
+## 0.3.0
+
+### Minor Changes
+
+- update version v0.3.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @echojs-ecosystem/persist@0.3.0
+  - @echojs-ecosystem/reactivity@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes