+++ packages/router/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @echojs-ecosystem/router
 
+## 0.3.0
+
+### Minor Changes
+
+- update version v0.3.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @echojs-ecosystem/hyperdom@0.3.0
+  - @echojs-ecosystem/reactivity@0.3.0
+  - @echojs-ecosystem/url-state@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes