+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @echojs-ecosystem/ui
 
+## 0.3.0
+
+### Minor Changes
+
+- update version v0.3.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @echojs-ecosystem/hyperdom@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes